pyproject.toml
[tool.poetry]
name = "obsrv"
version = "0.1.0"
description = ""
authors = ["Ravi Mula <ravismula@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pycryptodome = "^3.20.0"
psycopg2-binary = "^2.9.9"
pyyaml = "^6.0.1"
kafka-python = {git = "https://github.com/dpkp/kafka-python.git"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
coverage = "^7.4.3"
pyspark = "^3.5.1"
testcontainers = {extras = ["kafka", "postgres"], version = "^4.4.0"}
kafka-python = "^2.0.2"

[tool.poetry.group.batch.dependencies]
pyspark = "^3.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
